fix: navigate to cluster-settings page from clusters-menu
