+++ src/renderer/components/cluster-manager/cluster-view.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { observer } from "mobx-react";
 import { getMatchedCluster } from "./cluster-view.route";
 import { ClusterStatus } from "./cluster-status";
-import { hasLoadedView } from "./cluster-manager";
+import { hasLoadedView } from "./lens-views";
 
 @observer
 export class ClusterView extends React.Component {
